temporarily turn off 32-bit Windows SDE CI (#1406)
Windows SDE CI dimension seems to be failing since the original version
was deprecated. This updates the version to the latest one.
There are also some jobs that seem to "hang" and don't finish. This is
similar to what was happening with SDE+ASAN on Linux, so the same
workaround has been helping us kill some of these hanging test runs.
The workaround doesn't resolve all issues for Windows SDE 32-bit
however, some of the test runs fail to print "[ PASSED ]" at the end of
the test. There are some instances where the test seems to run out of
memory?
It seems like this hadn't been an issue prior to last week. The Windows
SDE 32-bit CI didn't have any hanging tests at well nor did any of these
issues exist.

I've cut a ticket to further investigate (P114059413). The CI dimension
is disabled for now to unblock PRs.